Refactor token tests and RSA JSON key tests
This allows new grant types to reuse test code for client_assertion and
creating test zones
